backend/requirements.txt
# Backend Python packages
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
pydantic==2.11.4
pydantic[email]
python-multipart==0.0.20
python-jose[cryptography]
bcrypt
groq
python-dotenv
